refactor(prism-codegen): share the port-tree walk and document the manifest
