+++ scripts/prism-codegen/score-cli.ts
@@ -1,11 +1,11 @@
 import * as fs from "fs/promises";
-import { readFileSync, readdirSync, statSync, existsSync } from "node:fs";
+import { readFileSync, existsSync } from "node:fs";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
 import { generateFromSource } from "./index.js";
 import { asyncMethodsForRailsFile } from "./async-source.js";
 import { TOPLEVEL } from "./codegen.js";
-import { TARGET_FILES, rubyAbsPath } from "./files.js";
+import { TARGET_FILES, TRAILS_AR_SRC, portTreeFiles, rubyAbsPath } from "./files.js";
 import { rubyFileToTs } from "./naming.js";
 import { scoreFile, indexPortTree, type ScoreEntry } from "./score.js";
 import {
@@ -30,28 +30,10 @@ import {
   staleSignOffs,
 } from "./signoff.js";
 
-const TRAILS_AR_SRC = "packages/activerecord/src";
 const HERE = path.dirname(fileURLToPath(import.meta.url));
 const API_COMPARE = path.join(HERE, "..", "api-compare");
 const BASELINE_PATH = path.join(HERE, "convergence-baseline.json");
 const SIGNOFF_PATH = path.join(HERE, "convergence-signoff.json");
-
-function portTreeFiles(): { path: string; source: string }[] {
-  const out: { path: string; source: string }[] = [];
-  const walk = (dir: string) => {
-    for (const e of readdirSync(dir)) {
-      const full = path.join(dir, e);
-      if (statSync(full).isDirectory()) {
-        if (e === "test-helpers" || e === "support") continue;
-        walk(full);
-      } else if (e.endsWith(".ts") && !e.endsWith(".test.ts")) {
-        out.push({ path: path.relative(TRAILS_AR_SRC, full), source: readFileSync(full, "utf8") });
-      }
-    }
-  };
-  walk(TRAILS_AR_SRC);
-  return out;
-}
 
 async function listJsonFiles(dir: string): Promise<string[]> {
   const out: string[] = [];
